refactor: unify agent scripts into single transparent interface
- Merge run_manager_developer_agents.py functionality into run_mcp_agent.py
- Use headless mode to automatically trigger manager-developer architecture
- Use interactive mode for single-agent with conversation loop
- Delete separate run_manager_developer_agents.py to eliminate code paths
- Update poe tasks to use unified script interface
- Preserve all existing CLI arguments and handoffs mechanism
- Update documentation to reflect unified interface
